removed QueuedRequestSenderFactory and its implementation as there is no need for requestId anymore. Using requestTicket now.
